Add plugin command support for `nix repl`
- Refactor the repl core into libexpr
- Use dependency injection via std::function to provide the completion
  functions from editline so we don't introduce extra dependencies for
  libexpr
- Add a RegisterReplCmd analogous to RegisterPrimOp for repl commands
- Refactor: get rid of the "ugly" global curRepl and replace it with a
  trick with closures on the nix side (that are effectively globals, but
  unique per repl user and thus not as ugly ;p)
- Rip out readline support since there appears to be no build system
  support for it and it is thus dead code
- Integration test this new plugin functionality
- Document it
