+++ doc/manual/command-ref/conf-file.xml
@@ -611,8 +611,8 @@ password <replaceable>my-password</replaceable>
         plugins may construct static instances of RegisterPrimOp to
         add new primops or constants to the expression language,
         RegisterStoreImplementation to add new store implementations,
-        RegisterCommand to add new subcommands to the
-        <literal>nix</literal> command, and RegisterSetting to add new
+        RegisterReplCmd to add new commands to the
+        <literal>nix</literal> REPL, and RegisterSetting to add new
         nix config settings. See the constructors for those types for
         more details.
       </para>
