+++ src/libutil/callable.hh
@@ -1,0 +1,30 @@
+#pragma once
+// from https://stackoverflow.com/a/48368508
+
+#include <type_traits>
+#include <functional>
+
+template <class F>
+struct lambda_traits : lambda_traits<decltype(&F::operator())>
+{ };
+
+template <typename F, typename R, typename... Args>
+struct lambda_traits<R(F::*)(Args...)> : lambda_traits<R(F::*)(Args...) const>
+{ };
+
+template <class F, class R, class... Args>
+struct lambda_traits<R(F::*)(Args...) const> {
+    using pointer = typename std::add_pointer<R(Args...)>::type;
+
+    static pointer cify(F&& f) {
+        static F fn = std::forward<F>(f);
+        return [](Args... args) {
+            return fn(std::forward<Args>(args)...);
+        };
+    }
+};
+
+template <class F>
+inline typename lambda_traits<F>::pointer cify(F&& f) {
+    return lambda_traits<F>::cify(std::forward<F>(f));
+}
